requirements.txt
tkinterweb
cryptography
markdown
google-generativeai
pygments